rename algorithm from matable to ols_cas21
